+++ src/Application/Configs/IceHawkConfig.php
@@ -58,12 +58,14 @@ final class IceHawkConfig implements ConfiguresIceHawk
 		$quotedBaseUri = preg_quote( $baseUrl, '!' );
 
 		$this->readRoutes = [
-			'^' . $quotedBaseUri . '/?$'                                                                                            => ServerSelectionRequestHandler::class,
-			'^' . $quotedBaseUri . '/server/(?<serverKey>\d+)/stats/?$'                                                             => ServerStatsRequestHandler::class,
-			'^' . $quotedBaseUri . '/server/(?<serverKey>\d+)(?:/database/(?<database>\d+))?/?$'                                    => ServerDetailsRequestHandler::class,
-			'^' . $quotedBaseUri . '/server/(?<serverKey>\d+)/database/(?<database>\d+)/keys/?$'                                    => AjaxSearchKeysRequestHandler::class,
-			'^' . $quotedBaseUri . '/server/(?<serverKey>\d+)/database/(?<database>\d+)/keys/(?<keyName>.+)/hash/(?<hashKey>.+)/?$' => AjaxKeyDetailsRequestHandler::class,
-			'^' . $quotedBaseUri . '/server/(?<serverKey>\d+)/database/(?<database>\d+)/keys/(?<keyName>.+)/?$'                     => AjaxKeyDetailsRequestHandler::class,
+			'^' . $quotedBaseUri . '/?$'                                                                        => ServerSelectionRequestHandler::class,
+			'^' . $quotedBaseUri . '/server/(?<serverKey>\d+)/stats/?$'                                         => ServerStatsRequestHandler::class,
+			'^' . $quotedBaseUri . '/server/(?<serverKey>\d+)(?:/database/(?<database>\d+))?/?$'                => ServerDetailsRequestHandler::class,
+			'^' . $quotedBaseUri . '/server/(?<serverKey>\d+)/database/(?<database>\d+)/keys/?$'                => AjaxSearchKeysRequestHandler::class,
+			'^'
+			. $quotedBaseUri
+			. '/server/(?<serverKey>\d+)/database/(?<database>\d+)/keys/(?<keyName>.+)/hash/(?<subKey>.+)/?$'   => AjaxKeyDetailsRequestHandler::class,
+			'^' . $quotedBaseUri . '/server/(?<serverKey>\d+)/database/(?<database>\d+)/keys/(?<keyName>.+)/?$' => AjaxKeyDetailsRequestHandler::class,
 		];
 	}
 }