+++ src/Application/ReadModel/Interfaces/ProvidesKeyData.php
@@ -1,0 +1,14 @@
+<?php declare(strict_types=1);
+
+namespace hollodotme\Readis\Application\ReadModel\Interfaces;
+
+interface ProvidesKeyData
+{
+	public function getKeyData() : string;
+
+	public function getRawKeyData() : string;
+
+	public function hasScore() : bool;
+
+	public function getScore() : ?float;
+}