+++ src/Infrastructure/Redis/ServerManager.php
@@ -2,7 +2,7 @@
 
 namespace hollodotme\Readis\Infrastructure\Redis;
 
-use hollodotme\Readis\Application\Interfaces\ProvidesKeyInformation;
+use hollodotme\Readis\Application\Interfaces\ProvidesKeyInfo;
 use hollodotme\Readis\Application\Interfaces\ProvidesSlowLogData;
 use hollodotme\Readis\Infrastructure\Interfaces\ProvidesConnectionData;
 use hollodotme\Readis\Infrastructure\Redis\DTO\KeyInfo;
@@ -99,7 +99,7 @@ final class ServerManager
 	 * @param string   $keyPattern
 	 * @param int|null $limit
 	 *
-	 * @return array|ProvidesKeyInformation[]
+	 * @return array|ProvidesKeyInfo[]
 	 * @throws ConnectionFailedException
 	 */
 	public function getKeyInfoObjects( string $keyPattern, ?int $limit ) : array
@@ -118,10 +118,10 @@ final class ServerManager
 	/**
 	 * @param string $key
 	 *
-	 * @return ProvidesKeyInformation
+	 * @return ProvidesKeyInfo
 	 * @throws ConnectionFailedException
 	 */
-	public function getKeyInfoObject( string $key ) : ProvidesKeyInformation
+	public function getKeyInfoObject( string $key ) : ProvidesKeyInfo
 	{
 		/** @noinspection PhpUndefinedMethodInspection */
 		[$type, $ttl] = $this->redis->multi()->type( $key )->pttl( $key )->exec();
@@ -130,13 +130,41 @@ final class ServerManager
 		{
 			/** @noinspection PhpUndefinedMethodInspection */
 			$subItems = $this->redis->hKeys( $key );
-		}
-		else
-		{
-			$subItems = [];
+
+			return new KeyInfo( $key, $type, $ttl, $subItems );
 		}
 
-		return new KeyInfo( $key, $type, $ttl, $subItems );
+		if ( $type === Redis::REDIS_LIST )
+		{
+			/** @noinspection PhpUndefinedMethodInspection */
+			$listLength = $this->redis->llen( $key );
+
+			/** @noinspection PhpUndefinedMethodInspection */
+			$subItems = $this->redis->lrange( $key, 0, $listLength - 1 );
+
+			return new KeyInfo( $key, $type, $ttl, $subItems );
+		}
+
+		if ( $type === Redis::REDIS_SET )
+		{
+			/** @noinspection PhpUndefinedMethodInspection */
+			$subItems = $this->redis->smembers( $key );
+
+			return new KeyInfo( $key, $type, $ttl, $subItems );
+		}
+
+		if ( $type === Redis::REDIS_ZSET )
+		{
+			/** @noinspection PhpUndefinedMethodInspection */
+			$setLength = $this->redis->zcard( $key );
+
+			/** @noinspection PhpUndefinedMethodInspection */
+			$subItems = $this->redis->zrange( $key, 0, $setLength - 1, true );
+
+			return new KeyInfo( $key, $type, $ttl, $subItems );
+		}
+
+		return new KeyInfo( $key, $type, $ttl, [] );
 	}
 
 	/**
@@ -162,5 +190,30 @@ final class ServerManager
 	{
 		/** @noinspection PhpUndefinedMethodInspection */
 		return $this->redis->hGet( $key, $hashKey );
+	}
+
+	/**
+	 * @param string $key
+	 *
+	 * @throws ConnectionFailedException
+	 * @return array
+	 */
+	public function getAllHashValues( string $key ) : array
+	{
+		/** @noinspection PhpUndefinedMethodInspection */
+		return (array)$this->redis->hGetAll( $key );
+	}
+
+	/**
+	 * @param string $key
+	 * @param int    $index
+	 *
+	 * @throws ConnectionFailedException
+	 * @return mixed
+	 */
+	public function getListValue( string $key, int $index )
+	{
+		/** @noinspection PhpUndefinedMethodInspection */
+		return $this->redis->lindex( $key, $index );
 	}
 }